Fix font regression in AdjustMMParams
Commit 2334660 changed the |dest_width| in CFX_Font::AdjustMMParams to
unsigned, but this means that dest_width - min_width becomes unsigned,
which is wrong because the subtraction could be negative. This CL fixes
this bug.

Bug: 845697

+++ core/fxge/cfx_font.cpp
@@ -535,7 +535,8 @@ void CFX_Font::AdjustMMParams(int glyph_index,
       FXFT_Free(m_Face, pMasters);
       return;
     }
-    int param = min_param + (max_param - min_param) * (dest_width - min_width) /
+    int param = min_param + (max_param - min_param) *
+                                (static_cast<int>(dest_width) - min_width) /
                                 (max_width - min_width);
     coords[1] = param;
   }
